[refactor] ChildVisit and PlaceVisit: detach from() to mapper class with extension function
